Changes to the ADC(1) and stieltjes_ap codes

+++ stieltjes_ap/simod.f90
@@ -2,12 +2,12 @@
 
     save
 
-    integer                              :: npoints,nord
-    integer, parameter                   :: ntrial=100
-    real*8, dimension(2)                 :: erange
-    real*8, dimension(:), allocatable    :: ener,osc
-    real*16, dimension(:), allocatable   :: si_e,si_f,si_osc,si_cosc1,&
-                                            si_cosc2
-    character(len=120)                   :: asiinp,aosc
+    integer                           :: npoints,nord
+    integer, parameter                :: ntrial=100
+    real*8, dimension(2)              :: erange
+    real*8, dimension(:), allocatable :: ener,osc
+    real*8, dimension(:), allocatable :: si_e,si_f,si_osc,si_cosc1,&
+                                         si_cosc2
+    character(len=120)                :: asiinp,aosc
 
   end module simod
